Revert "close_header_menu_when_click_anywhere_except_for_icon" because pushed wrong branch
This reverts commit 96cc08fd412b1977bdcfa98347c6964f1463b178.
