+++ app/javascript/components/TheHeader.vue
@@ -10,6 +10,7 @@
             <img
               :src="logo_src"
               width="120px"
+              @click="closeMenu"
             >
           </router-link>
           <div class="item">
@@ -20,10 +21,7 @@
             >
               Twitter認証によるユーザー登録
             </a>
-            <button
-              @click="openHeaderMenu"
-              v-click-outside="handleCloseHeaderMenu"
-            >
+            <button @click="isOpen = !isOpen">
               <img
                 v-if="currentUser"
                 :src="currentUser.image"
@@ -32,36 +30,56 @@
             </button>
           </div>
         </div>
-        <HeaderMenu
-          :is-visible-header-menu="isVisibleHeaderMenu"
-          @close-menu="handleCloseHeaderMenu"
-        />
+        <div
+          v-if="currentUser"
+          :class="isOpen ? 'flex' : 'hidden'"
+          class="flex-col mt-3 justify-between"
+        >
+          <button
+            class="p-2 lg:px-4 text-2xl text-center border border-transparent rounded hover:bg-red-100 hover:text-red-700 transition-colors duration-300"
+          >
+            <router-link :to="`/users/${currentUser.screen_name}`">
+              マイページ
+            </router-link>
+          </button>
+          <button
+            class="p-2 lg:px-4 text-2xl text-center border border-transparent rounded hover:bg-red-100 hover:text-red-700 transition-colors duration-300"
+          >
+            <router-link to="/previous">
+              過去の診断結果
+            </router-link>
+          </button>
+          <button
+            class="p-2 lg:px-4 text-2xl text-center border border-transparent rounded hover:bg-red-100 hover:text-red-700 transition-colors duration-300"
+            @click="logout"
+          >
+            ログアウト
+          </button>
+        </div>
       </div>
     </nav>
   </header>
 </template>
 
 <script>
-import HeaderMenu from "./HeaderMenu"
 import { mapGetters　} from "vuex"
-import ClickOutside from "vue-click-outside"
 
 export default {
-  directives: {
-    ClickOutside
-  },
+  name: 'TheHeader',
   components: {
-    HeaderMenu
   },
   data() {
     return {
-      isVisibleHeaderMenu: false,
+      isOpen: false,
     }
   },
   computed: {
     ...mapGetters(
       'users', ['currentUser' , 'isAuthenticatedUser']
     ),
+    currentPath() {
+      return this.$route.path
+    },
     logo_src() {
       return require("../../../public/images/topLogo.png")
     },
@@ -83,11 +101,13 @@ export default {
         console.log(err.response)
       }
     },
-    openHeaderMenu(){
-      this.isVisibleHeaderMenu = true;
+    closeMenu(){
+      this.isOpen = false;
     },
-    handleCloseHeaderMenu(){
-      this.isVisibleHeaderMenu = false;
+  },
+  watch: {
+    currentPath: function() {
+      this.closeMenu()
     }
   },
 }
