+++ package.json
@@ -20,7 +20,6 @@
     "vee-validate": "^3.4.13",
     "vue": "^2.6.14",
     "vue-chartjs": "3.5.1",
-    "vue-click-outside": "^1.1.0",
     "vue-gtag": "^1.16.1",
     "vue-loader": "15.9.2",
     "vue-router": "^3.5.2",
